[move-prover] Specification of association.move, partial specs for libra.move
Both are partially completed, but association.move is more complete
than libra.move.  I am making incomplete use of doc gen in association.move.

In association.move, there is a *feature request* that we should discuss.

There is one property that doesn't hold (root is a member of the
association). I need to discuss with Tim.

Closes: #3991
Approved by: wrwg