[move-prover] Generalized quantifiers and types-as-values
This PR adds a lot of new fire power (also to shoot in your foot) for specifications with a few extensions:

1. We now support types as values. You can write `type<T>()` to construct a value which represents type T. The only operation available on type values
   is equality. The type of types is called `type`, so one can e.g. do
   ```
   define helper(t: type, addr: address): bool { exists<R<t>>(addr) }
   ```
   Notice the use of the type value `t` in the instantiation of `exists`.

2. The quantification over addresses David added recently has beeen generalized to arbitrary types. You can write
   `all(domain<address>, |addr| P(addr))` to quantify over addresses, for example. This also works for types using `domain<type>()`.

This was technically relative simple because the boogie encoding already deals with types as values.

The example in the test case `type_values.move` demonstrates that this mechanism can be used to express resource invariants, in the presence of generic types,
as module invariants, which -- surprisingly -- seems to work well. Similarly, this should allow to write schemas which can apply to all functions in a module
and do not require type parameters (resp. use old-style module invariants directly).

The syntax we have choosen for quantifiers may not be totally adequate any longer, as it was oriented towards bounded quantification over vectors or ranges. We may want to change it in the near future.

Closes: #3980
Approved by: DavidLDill