[backup] StateSnapshotBackupController
Refactoring to to extract backup logic to "controller"s. From this on
the CLI main entry point is supposed to construct shared facilities
among controllers (like the storage and the backup-service client),
pass parameters from the command line or config file transparently to
the controllers and run them.

Closes: #4083
Approved by: wqfish