[cluster-test] Introduce Instance::{start, stop, util_cmd}
This diff introduces convenience methods on the Instance.

They only available in k8s right now, and calls to appropriate methods in ClusterSwarmKube.

This allows to have Instance as a self-sufficient entity and not to pass reference to swarm to every place where we need to manipulate instance state

Also removes legacy(and no longer working) Instance:run_cmd and family

Closes: #4364
Approved by: ankushagarwal