[language] adjust VM counters for transaction timing
The six different performance timers that we've been using for the
VM are not actually very useful. Most of the time is spent waiting
for other parts of the system, so the breakdown into different parts
of the VM processing just confuses things. Drop down to use only
two timers: overall user transaction time and overall validation
time.

Closes: #5105
Approved by: dariorussi