[sccache] use checksum of toolchains in cache key
Both the sccache binary and the compilation cache are cached between
jobs. Previously, the binary was indexed by sccache version, while the
compilation cache was indexed by a combination of job type, toolchain,
and date. The discrepncy should be corrected.

Both the binary and the compilation cache use checksum of the same
set of toolchain version files in the index key. The compilation
adds the date code in addition to the key to keep the cache fresh for
the day. After development stablize, the date code can be removed.