[state sync] remove gossip discovery dependency from integration_tests
This diff is made in preparation for removing the gossip discovery
protocol from network.

Previously, this state sync integration test framework would add the
previously generated peer as a seed peer and let gossip discovery do
the rest of the work to learn all other peer addresses and then connect
them. With this diff, we "pre-bind" the `/memory/<port>` addresses and
just build the seed peer set from the beginning.

On a slightly tangential note, it seems strange to me that this
framework puts all peers (both validators and full nodes) on the same
network (when it's not mocking the network interface), which doesn't
accurately reflect reality. Something to change later I suppose.