[prover] fine-grained read and write set analysis
The analysis introduced in this PR seeks to overapproximate the set of *concrete access paths* that may be read/written by a given procedure. An access path is a canonical name for a location in memory; some examples of are:

`0x7/M::T/f` (i.e., the field `f` of the `M::T` resource stored at address `0x7`

`Formal(0)/[2]` (i.e., the value stored at index 2 of the array bound the 0th formal of the current procedure)

The analysis uses an abstraction of access paths with the following components

- A *root*, which is either an abstract address or a local
- Zero or more *offsets*, where an offset is a field, an unknown vector index, or an abstract struct type

Abstract addresses are a set containing constants and abstract access paths read from the environment (or "footprint"). For example, in the following code:

```
struct S { f: u64 }

fun foo(x: S) {
  let a = if (*) { 0x1 } else { *&x.f }
   ...
}
```

we will represent `a` as `{ 0x1, Footprint(x/f) }`

The obvious approach to abstracting a set of concrete paths is using a set of abstract paths. The analysis does exactly this, but using an implementation that avoids redundant representations of the same memory: an *access path trie* where roots nodes are access path roots and each internal node is an access path offset. Each node is (optionally) associated with abstract value of a generic type `T`.

The read/write set analysis state consists of two such tries: an `AccessPathTrie<Access>` to track the accesses to global state, and an `AccessPathTrie<AbsAddr>` to track the set of address values at each program point. An access is either Borrow, Read, Write, or ReadWriteBorrow (top value). The Borrow value is not strictly needed, but it helps for debugging.

The read/write set analysis is a compositional analysis that starts from the leaves of the call graph and analyzes each procedure once. The result is a summary of the abstract paths read/written by each procedure and the value(s) returned by the procedure.

When the analysis encounters a call, it fetches the summary for the callee and applies it to the current state. This logic (implemented in `apply_summary`) is by far the most complex part of the analysis. There are three steps:

1. Substitute footprint values in the callee summary with their values in the caller state (including both actuals and values read from globals)
2. Bind return values in the callee summary to the return variables in the caller state
3. Join caller accesses and callee accesses

There is a lot left to do before this analysis is useful/mature:
- Add many, many more tests. There are a few in this PR, and I have run on the Diem Framework/eyeballed several scripts to make sure the results look reasonable. But there are a ton of corner cases to test here.
- Expand the analysis from procedures to transactions: we must substitute in the concrete values provided in a transaction into the summary + (possibly) read some values from the current on-chain state. In addition, this analysis must take the prologue and epilogue into account.
- Make it possible to build a `GlobalEnv` from bytecode
- Address the termination issues by adding widening. Today, the analysis will hang if you write "address linked list" code that creates access paths of unknown length. We don't have any code like this in the framework, but it is important to handle this case in an open system
- Improve the precision of the analysis by adding guards to each access node. This would help us with the "all txes conflict because of gas issue", at least in the case where the price is nonzero.

Closes: #7330