[language] Move crate language/move-prover/spec-lang to language/move-model
The spec-lang crate has become a general tool for reflecting over move code, used by tools other than the prover. This PR is a 1st step to reflect this correctly and move it out of the prover. In subsequent PRs, some tools currently in move-prover will be moved out there as well and decoupled from the prover command line driver.

The chosen name is `move-model`, reflecting that this crate's modules provide a model of the Move language which is on a higher abstraction level than file format. It nevertheless contains also the logic for type checking the spec language, which may itself at some point be moved to a different place.

Closes: #7068