[read/write set] handle addresses flowing into/out of structs via pack and unpack
Previously, the read/write set analysis treated packs and unpacks as no-ops. However, this is incorrect
because the analysis must track addresses flowing into and out of structs.

This PR adds support for pack and unpack in the natural way by:
- adding each packed field of type struct or address to the locals trie
- adding each unpacked field of type struct or address to the locals trie

In the process of supporting pack/unpack, I uncovered a bug in `AccessPath::substitute_footprint` (see the `use_results` function in `pack_unpack.move` and the TODO in access_path.rs). Fixing this bug will require a fairly significant refactoring, so I will defer it to a subsequent PR.

Closes: #8658