[move-prover] Remove last usage of `into_module`
The method `CompiledScript::into_module` is deprecated and no longer
used inside of the Move VM. Unfortunately, its usage within the prover
is more difficult to expunge, and would involve rewriting several parts
to operate upon `BinaryIndexedView`, a facade over either a module or
script.

For now, instead, move the logic in `CompiledScript::into_module` into
a private function in the prover, and delete the deprecated method.

Closes: #8655