[move] Remove unused generic `Location` parameters
Many types in the Move codebase are defined to operate upon a generic
"source location" type, in order to operate on both `OwnedLoc` (a source
location type, now deleted, that stored its `file` field as a `String`),
and on `Loc` (a source location type that used to store its `file`
field as a `&str` reference, and now stores it as a globally interned
string `Symbol`). `OwnedLoc` has since been removed, and so these
generic types can be made into simpler, concrete ones.

In addition, helper functions implemented to map between types that use
`OwnedLoc` and those that use `Loc` can also be deleted.

Closes: #9022
Closes: #10652