json-rpc: implement missing MockDB methods to fix fuzzer
Continuous fuzzing discovered a panic in the json-rpc fuzzer where one
of the fuzzers managed to discover `get_accumulator_consistency_proof`
and called it. The fuzzer, however, uses a MockDB rather than a real
DiemDB. The MockDB didn't have `get_accumulator_consistency_proof`
implemented and so the fuzz test panicked with `unimplemented!()`.
This means the fuzz panic is a false positive, but something we should
fix nonetheless so it can discover more useful stuff.

To fix the problem, this PR back-fills all of the missing DB methods
called by the json-rpc service with dummy implementations so the fuzzer
won't panic if it discovers other methods. We also add the rest of the
json-rpc methods (except `get_resources`) to the `test_method_fuzzer`
test so we can surface these `unimplemented!()` panics sooner in CI.

Closes: #9128